requirements.txt
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
PyQt5==5.15.2
PyQt5-sip==12.8.1
PyQt5-stubs==5.14.2.2
